[refactoring] moved sensu api to api/process
