+++ bin/sensu-api
@@ -1,10 +1,10 @@
 #!/usr/bin/env ruby
 
-unless $:.include?(File.dirname(__FILE__) + '/../lib/')
-  $: << File.dirname(__FILE__) + '/../lib'
+unless $:.include?(File.dirname(__FILE__) + "/../lib/")
+  $: << File.dirname(__FILE__) + "/../lib"
 end
 
-require 'sensu/api'
+require "sensu/api/process"
 
 options = Sensu::CLI.read
-Sensu::API.run(options)
+Sensu::API::Process.run(options)
