+++ bin/sensu-server
@@ -1,10 +1,10 @@
 #!/usr/bin/env ruby
 
-unless $:.include?(File.dirname(__FILE__) + '/../lib/')
-  $: << File.dirname(__FILE__) + '/../lib'
+unless $:.include?(File.dirname(__FILE__) + "/../lib/")
+  $: << File.dirname(__FILE__) + "/../lib"
 end
 
-require 'sensu/server/process'
+require "sensu/server/process"
 
 options = Sensu::CLI.read
 Sensu::Server::Process.run(options)
